64-bit port, find socket libraries.
git-svn-id: svn://openpgp.nominet.org.uk/openpgpsdk/trunk@192 228a932f-eeec-0310-9813-fb354622df21

+++ include/packet.h
@@ -638,7 +638,7 @@ typedef struct
 
 typedef struct
     {
-    size_t			length;
+    unsigned			length;
     unsigned char		data[8192];
     } ops_literal_data_body_t;
